refactor(hgraph): create hgraph directory and split cpp files
