+++ src/algorithm/hgraph/hgraph.h
@@ -19,6 +19,7 @@
 #include <shared_mutex>
 #include <string>
 
+#include "../inner_index_interface.h"
 #include "common.h"
 #include "datacell/attribute_inverted_interface.h"
 #include "datacell/flatten_interface.h"
@@ -34,7 +35,6 @@
 #include "index/iterator_filter.h"
 #include "index_common_param.h"
 #include "index_feature_list.h"
-#include "inner_index_interface.h"
 #include "typing.h"
 #include "utils/lock_strategy.h"
 #include "utils/util_functions.h"
