+++ src/algorithm/hgraph/hgraph_parameter.h
@@ -17,9 +17,9 @@
 
 #include <limits>
 
+#include "../index_search_parameter.h"
+#include "../inner_index_parameter.h"
 #include "data_type.h"
-#include "index_search_parameter.h"
-#include "inner_index_parameter.h"
 #include "utils/pointer_define.h"
 #include "vsag/constants.h"
