+++ src/algorithm/pyramid_zparameters.h
@@ -17,7 +17,7 @@
 
 #include <functional>
 
-#include "algorithm/hgraph_parameter.h"
+#include "algorithm/hgraph/hgraph_parameter.h"
 #include "algorithm/index_search_parameter.h"
 #include "datacell/flatten_interface.h"
 #include "datacell/graph_datacell_parameter.h"
