+++ src/index/index_impl_test.cpp
@@ -17,7 +17,7 @@
 
 #include <sstream>
 
-#include "algorithm/hgraph.h"
+#include "algorithm/hgraph/hgraph.h"
 #include "index_common_param.h"
 #include "unittest.h"
 #include "vsag/engine.h"
