+++ tools/analyze_index/analyze_index.cpp
@@ -21,7 +21,7 @@
 #include <iostream>
 #include <memory>
 
-#include "algorithm/hgraph.h"
+#include "algorithm/hgraph/hgraph.h"
 #include "algorithm/ivf.h"
 #include "algorithm/pyramid.h"
 #include "algorithm/pyramid_zparameters.h"
